prestiti/requirements.txt
Flask
mysql-connector-python
couchdb
